Add search and course page
Added a course and search page together with a bunch of view models that
they need. The way the pivot works in the course page unfortunately makes
blend lose its marbles. This will make changes to these pages more
tedious. It is possible though to work around, and that'll be added later.

Page navigation is not very nice. It has to be fixed later.