Move oodi and noppa buttons from overview pivot to appbar
Added new buttons to the app bar, one for opening the course in noppa
and one for opening it in oodi. Changed the pin icon to a star, and
added a crappy oodi icon. It needs to be made better and in a flatter
style (maybe no perespective).

Fixed bug where after pinning a course, the pin button remains disabled.